% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoElnet_function.R
\name{calculate_perf_rrace}
\alias{calculate_perf_rrace}
\title{calculate_perf_rrace}
\usage{
calculate_perf_rrace(data, weight, prop, sr)
}
\arguments{
\item{data}{input data (matrix), which consists of predictor scores, performance score (column label: "Perf") and race dummy variable (column label: "Race_dummy")}

\item{weight}{predictor weights}

\item{prop}{proportion of minority applicants = (number of minority applicants)/(number of all applicants)}

\item{sr}{selection ratio = (number of hires) / (number of applicants)}
}
\value{
Perf job performance criterion validity (i.e., correlation between job performance score and predictor weighted composite score)

rrace correlation between race dummy variable and predictor composite
}
\description{
Function for calculating job performance criterion validity and r_race using predictor weights
}
