% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoElnet.R
\name{gen_data}
\alias{gen_data}
\title{gen_data}
\usage{
gen_data(prop, sr, n, D, R)
}
\arguments{
\item{prop}{Proportion of minority applicants in full applicant pool}

\item{sr}{Selection ratio}

\item{D}{Subgroup difference: standardized mean differences between minority and majority subgroups (i.e., majority - minority), on each predictor and criterion}

\item{R}{Correlation matrix: Criterion predictor inter-correlation matrix (in applicant pool); format: Predictor_1, ..., Predictor_n, Criterion}
}
\value{
data_out Generated sample data
}
\description{
Function to generate sample (used in cv.ParetoElnet)
}
\examples{
# Specify inputs
# (1) Proportion of minority applicants (prop) = (# of minority applicants)/(total # of applicants)
prop <- 1/4
# (2) Selection ratio (sr) = (# of selected applicants)/(total # of applicants)
sr <- 0.10
# (3) Sample size
n <- 100
# (4) Subgroup differences (d): standardized mean differences between minority and majority subgroups (i.e., majority - minority), on each predictor 
D <- c(1.00, 0.23, 0.09, 0.33, .30)
# (5) Correlation matrix (R) = criterion & predictor inter-correlation matrix (in applicant pool)
# Format: Predictor_1, ..., Predictor_n, Criterion
R <- matrix(c(1, .24, .00, .19, .30,
              .24, 1, .12, .16, .30,
              .00, .12, 1, .51, .18,
              .19, .16, .51, 1, .28,
              .30, .30, .18, .28, 1),
            (length(d)+1),(length(d)+1))
# Generate data
data_out = gen_data(prop, sr, n, D, R)

}
