% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoFunction.r
\name{myCon_eq}
\alias{myCon_eq}
\title{myCon_eq}
\usage{
myCon_eq(x)
}
\arguments{
\item{x}{Input predictor weight vector}
}
\value{
Equal constraint condition for use in NBI()
}
\description{
Support function, defines equal constraint condition
}
