% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoElnet_function.R, R/ParetoR_function.R
\name{myCon_ineq}
\alias{myCon_ineq}
\alias{myCon_ineq}
\title{myCon_ineq}
\usage{
myCon_ineq(x)

myCon_ineq(x)
}
\arguments{
\item{x}{Input predictor weight vector}

\item{x}{Input predictor weight vector}
}
\value{
Inequal constraint condition for use in NBI()

Inequal constraint condition for use in NBI()
}
\description{
Support function, defines inequal constraint condition

Support function, defines inequal constraint condition
}
