% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoFunction_fromShiny.R
\name{myFM}
\alias{myFM}
\title{myFM}
\usage{
myFM(x)
}
\arguments{
\item{x}{Input predictor weight vector}
}
\value{
f Criterion vector
}
\description{
Supporting function, defines criterion space
}
