% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoElnet_function.R
\name{myLinCom_Elnet}
\alias{myLinCom_Elnet}
\title{myLinCom_Elnet}
\usage{
myLinCom_Elnet(x)
}
\arguments{
\item{x}{Input predictor weight vector}
}
\value{
f Criterion vector
}
\description{
Support function
}
