% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoElnet_function.R
\name{NBI_Elnet}
\alias{NBI_Elnet}
\title{NBI Main Function}
\usage{
NBI_Elnet(X0, Spac, Fnum, VLB = vector(), VUB = vector(), TolX = 1e-04,
  TolF = 1e-04, TolCon = 1e-07, graph = TRUE, display_solution = TRUE)
}
\arguments{
\item{X0}{Initial input for preditor weight vector}

\item{Spac}{Number of Pareto spaces (i.e., number of Pareto points minus one)}

\item{Fnum}{Number of criterions}

\item{VLB}{Lower boundary for weight vector estimation}

\item{VUB}{Upper boundary for weight vector estimation}

\item{TolX}{Tolerance index for estimating weight vector, default is 1e-4}

\item{TolF}{Tolerance index for estimating criterion, default is 1e-4}

\item{TolCon}{Tolerance index for constraint conditions, default is 1e-7}

\item{graph}{If TRUE, plots will be generated for Pareto-optimal curve and predictor weights}

\item{display_solution}{If TRUE, Pareto-optimal solution will be displayed}
}
\value{
Pareto-Optimal solutions
}
\description{
Main function for obtaining pareto-optimal solution via normal boundary intersection.
}
