% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoElnet.R
\name{ParetoElnet}
\alias{ParetoElnet}
\title{ParetoElnet}
\usage{
ParetoElnet(prop, sr, d, R, Spac = 21, graph = TRUE,
  display_solution = TRUE)
}
\arguments{
\item{prop}{Proportion of minority applicants in full applicant pool}

\item{sr}{Selection ratio}

\item{d}{Subgroup difference}

\item{R}{Correlation matrix}

\item{Spac}{Number of Pareto points (e.g., 21)}

\item{graph}{If TRUE, plots will be generated for Pareto-optimal curve and predictor weights}

\item{display_solution}{If TRUE, Pareto-optimal solutions will be printed}
}
\value{
out Pareto-Optimal solution with criterion values (Criterion) and predictor weights (ParetoWeights)
}
\description{
Command function to run Regularized Pareto-Optimal algorithm.
}
\examples{
# Specify inputs
# (1) Proportion of minority applicants (prop) = (# of minority applicants)/(total # of applicants)
prop <- 1/4
# (2) Selection ratio (sr) = (# of selected applicants)/(total # of applicants)
sr <- 0.10
# (3) Subgroup differences (d): standardized mean differences between minority and majority subgroups (i.e., majority - minority), on each predictor (in applicant pool)
d <- c(1.00, 0.23, 0.09, 0.33)
# (4) Correlation matrix (R) = criterion & predictor inter-correlation matrix (in applicant pool)
# Format: Predictor_1, ..., Predictor_n, Criterion
R <- matrix(c(1, .24, .00, .19, .30,
              .24, 1, .12, .16, .30,
              .00, .12, 1, .51, .18,
              .19, .16, .51, 1, .28,
              .30, .30, .18, .28, 1),
            (length(d)+1),(length(d)+1))
# Fit Regularized Pareto-optimal model
out = ParetoElnet(prop, sr, d, R)

}
