% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoR.R
\name{ParetoR}
\alias{ParetoR}
\title{ParetoR}
\usage{
ParetoR(prop, sr, d, R, graph = TRUE, display_solution = TRUE)
}
\arguments{
\item{prop}{Proportion of minority applicants in full applicant pool}

\item{sr}{Selection ratio}

\item{d}{Subgroup difference}

\item{R}{Correlation matrix}

\item{graph}{If TRUE, plots will be generated for Pareto-optimal curve and predictor weights}

\item{display_solution}{If TRUE, Pareto-optimal solution will be displayed}
}
\value{
out Pareto-Optimal solution with criterion values (Criterion) and predictor weights (ParetoWeights)
}
\description{
Command function to run Pareto-Optimal algorithm.
}
\examples{
# Specify inputs
# (1) Proportion of minority applicants (prop) = (# of minority applicants)/(total # of applicants)
## Example:
prop <- 1/4
# (2) Selection ratio (sr) = (# of selected applicants)/(total # of applicants)
## Example:
sr <- 0.10
# (3) Subgroup differences (d): standardized mean differences between minority and majority subgroups (i.e., majority - minority), on each predictor (in applicant pool)
## Example:
d <- c(1.00, 0.23, 0.09, 0.33)
# (4) Correlation matrix (R) = criterion & predictor inter-correlation matrix (in applicant pool)
## Example:
# Format: Predictor_1, ..., Predictor_n, Criterion
R <- matrix(c(1, .24, .00, .19, .30,
              .24, 1, .12, .16, .30,
              .00, .12, 1, .51, .18,
              .19, .16, .51, 1, .28,
              .30, .30, .18, .28, 1),
            (length(d)+1),(length(d)+1))
# Fit Pareto-optimal model
out = ParetoR(prop, sr, d, R)
}
