% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoShrinkage.R
\name{ParetoShrinkage}
\alias{ParetoShrinkage}
\title{ParetoShrinkage function}
\usage{
ParetoShrinkage(Ncal, numpred, p_pareto, R_perf_cal, R_race_cal,
  use = "Wherry")
}
\arguments{
\item{Ncal}{Calibration sample size}

\item{numpred}{number of predictors}

\item{p_pareto}{number of Pareto-optimal points (e.g., 21)}

\item{R_perf_cal}{Vector of calibration sample job performance validity}

\item{R_race_cal}{Vector of calibration sample race bivariate correlation [i.e., correlation between race dummy variable (0-minority, 1-majority) and predictor composite score]}

\item{use}{Shrinkage formula to be used at first step when generalizing to population. Options for "use": "Wherry" - Wherry (1931) shrinkage formula; "OlkinPratt" - simplified Olkin & Pratt (1958) formula (see Cattin, 1980, p. 409), "OlkinPratt Hypergeo" - Olkin & Pratt (1958) formula with hypergeometric function; "Claudy" - Claudy (1978) formula. Default is "Wherry".}
}
\value{
out formula-adjusted Pareto-optimal solution
}
\description{
Estimate shrunken Pareto-optimal solution based on Pareto-optimal shrinkage formulae
}
