% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoFunction.R
\name{plotPareto}
\alias{plotPareto}
\title{plotPareto}
\usage{
plotPareto(Pareto_Fmat, Pareto_Xmat)
}
\arguments{
\item{Pareto_Fmat}{Pareto-Optimal criterion solution}

\item{Pareto_Xmat}{Pareto-Optimal predictor weight solution}
}
\value{
Plot of Pareto-optimal curve and plot of predictor weights
}
\description{
Function for plotting Pareto-optimal curve and predictor weights
}

