% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoShrinkage.R
\name{R2_Browne}
\alias{R2_Browne}
\title{R2_Browne function}
\usage{
R2_Browne(N, p, R2, use = "Wherry")
}
\arguments{
\item{N}{Sample size}

\item{p}{number of predictors}

\item{R2}{R-squared}

\item{use}{Shrinkage formula to be used at first step when generalizing to population. Options for "use": "Wherry" - Wherry (1931) shrinkage formula; "OlkinPratt" - simplified Olkin & Pratt (1958) formula (see Cattin, 1980, p. 409), "OlkinPratt Hypergeo" - Olkin & Pratt (1958) formula with hypergeometric function; "Claudy" - Claudy (1978) formula. Default is "Wherry".}
}
\value{
R2_B formula-adjusted R2 based on Browne (1958) shrinkage formula
}
\description{
Estimate shrunken R2 based on Browne (1975) formula
}
\examples{
# (1) Sample size
N <- 100
# (2) Number of predictors
p <- 5
# (3) R2 R-squared
R2 <- 0.30
# Estimate shrunken R2
R2_Browne(N = N, p = p, R2 = R2, use = "Wherry")
}
