% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoAdj.R
\name{R2_CB}
\alias{R2_CB}
\title{R2_CB function}
\usage{
R2_CB(N, p, R2)
}
\arguments{
\item{N}{Sample size}

\item{p}{number of predictors}

\item{R2}{R-squared}
}
\value{
R2_B formula-adjusted R2 based on Browne (1958) shrinkage formula
}
\description{
Estimate shrunken R2 based on Browne (1975) formula
}
\examples{
# (1) Sample size
N <- 100
# (2) Number of predictors
p <- 5
# (3) R2 R-squared
R2 <- 0.30
# Estimate shrunken R2
R2_CB(N = N, p = p, R2 = R2)
}
