% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoShrinkage.R
\name{R2_Claudy}
\alias{R2_Claudy}
\title{R2_Claudy function}
\usage{
R2_Claudy(N, p, R2)
}
\arguments{
\item{N}{Sample size}

\item{p}{number of predictors}

\item{R2}{R-squared}
}
\value{
R2_C formula-adjusted R2 based on Claudy (1978) shrinkage formula
}
\description{
Estimate shrunken R2 based on Claudy (1978) formula
}
\examples{
# (1) Sample size
N <- 100
# (2) Number of predictors
p <- 5
# (3) R2 R-squared
R2 <- 0.30
# Estimate shrunken R2
R2_Claudy(N = N, p = p, R2 = R2)
}
