% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoShrinkage.R
\name{R2_OP}
\alias{R2_OP}
\title{R2_OP function}
\usage{
R2_OP(N, p, R2)
}
\arguments{
\item{N}{Sample size}

\item{p}{number of predictors}

\item{R2}{R-squared}
}
\value{
R2_OP formula-adjusted R2 based on simplified Olkin & Pratt (1958) shrinkage formula (see Cattin, 1980, p. 409)
}
\description{
Estimate shrunken R2 based on simplified Olkin & Pratt (1958) formula (see Cattin, 1980, p. 409)
}
\examples{
# (1) Sample size
N <- 100
# (2) Number of predictors
p <- 5
# (3) R2 R-squared
R2 <- 0.30
# Estimate shrunken R2
R2_OP(N = N, p = p, R2 = R2)
}
