% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoShrinkage.R
\name{R2_OPh}
\alias{R2_OPh}
\title{R2_OPh function}
\usage{
R2_OPh(N, p, R2)
}
\arguments{
\item{N}{Sample size}

\item{p}{number of predictors}

\item{R2}{R-squared}
}
\value{
R2_OPh formula-adjusted R2 based on Olkin & Pratt (1958) shrinkage formula with hypergeometric function
}
\description{
Estimate shrunken R2 based on Olkin & Pratt (1958) formula with hypergeometric function
}
\examples{
# (1) Sample size
N <- 100
# (2) Number of predictors
p <- 5
# (3) R2 R-squared
R2 <- 0.30
# Estimate shrunken R2
R2_OPh(N = N, p = p, R2 = R2)
}
