% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoFunction.R
\name{WeightsFun}
\alias{WeightsFun}
\title{WeightsFun}
\usage{
WeightsFun(n, k)
}
\arguments{
\item{n}{Number of objects (i.e., number of predictor and criterion)}

\item{k}{Number of Pareto points}
}
\value{
Weights All possible weights for NBI subproblem
}
\description{
Support function, generates all possible weights for NBI subproblems
}

