% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoElnet_function.R, R/ParetoR_function.R
\name{assert_col_vec}
\alias{assert_col_vec}
\alias{assert_col_vec}
\title{assert_col_vec}
\usage{
assert_col_vec(v)

assert_col_vec(v)
}
\arguments{
\item{v}{Intermediate variable v}

\item{v}{Intermediate variable v}
}
\value{
Refined variable v

Refined variable v
}
\description{
Support function, refines intermediate variable for use in NBI()

Support function, refines intermediate variable for use in NBI()
}
