% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoElnet_function.R, R/ParetoR_function.R
\name{myCon_eq}
\alias{myCon_eq}
\alias{myCon_eq}
\title{myCon_eq}
\usage{
myCon_eq(x)

myCon_eq(x)
}
\arguments{
\item{x}{Input predictor weight vector}

\item{x}{Input predictor weight vector}
}
\value{
Equal constraint condition for use in NBI()

Equal constraint condition for use in NBI()
}
\description{
Support function, defines equal constraint condition

Support function, defines equal constraint condition
}
