% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoElnet_function.R, R/ParetoR_function.R
\name{myTCon_eq}
\alias{myTCon_eq}
\alias{myTCon_eq}
\title{myTCon_eq}
\usage{
myTCon_eq(x_t)

myTCon_eq(x_t)
}
\arguments{
\item{x_t}{Temporary input weight vector}

\item{x_t}{Temporary input weight vector}
}
\value{
ceq Temporary constraint condition

ceq Temporary constraint condition
}
\description{
Support function, define constraint condition for intermediate step in NBI()

Support function, define constraint condition for intermediate step in NBI()
}
