% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pareto_command_file.R
\name{ParetoR}
\alias{ParetoR}
\title{ParetoR}
\usage{
ParetoR(prop, sr, d, R)
}
\arguments{
\item{prop}{Proportion of minority applicants in full applicant pool}

\item{sr}{Selection ratio}

\item{d}{Subgroup difference}

\item{R}{Correlation matrix}
}
\value{
out Pareto-Optimal solution
}
\description{
Command function to run Pareto-Optimal algorithm.
}

