% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoElnet_function.R, R/ParetoR_function.R
\name{plotPareto}
\alias{plotPareto}
\alias{plotPareto}
\title{plotPareto}
\usage{
plotPareto(Pareto_Fmat, Pareto_Xmat)

plotPareto(Pareto_Fmat, Pareto_Xmat)
}
\arguments{
\item{CriterionOutput}{Pareto-Optimal criterion solution}

\item{ParetoWeights}{Pareto-Optimal predictor weight solution}

\item{CriterionOutput}{Pareto-Optimal criterion solution}

\item{ParetoWeights}{Pareto-Optimal predictor weight solution}
}
\value{
Plot of Pareto-optimal curve and plot of predictor weights

Plot of Pareto-optimal curve and plot of predictor weights
}
\description{
Function for plotting Pareto-optimal curve and predictor weights

Function for plotting Pareto-optimal curve and predictor weights
}
