% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoElnet_function.R, R/ParetoR_function.R
\name{WeightsFun}
\alias{WeightsFun}
\alias{WeightsFun}
\title{WeightsFun}
\usage{
WeightsFun(n, k)

WeightsFun(n, k)
}
\arguments{
\item{n}{Number of objects (i.e., number of predictor and criterion)}

\item{k}{Number of Pareto points}

\item{n}{Number of objects (i.e., number of predictor and criterion)}

\item{k}{Number of Pareto points}
}
\value{
Weights All possible weights for NBI subproblem

Weights All possible weights for NBI subproblem
}
\description{
Support function, generates all possible weights for NBI subproblems

Support function, generates all possible weights for NBI subproblems
}
